extracted products row view from the product view
